+++ keras_nlp/models/xlnet/xlnet_encoder.py
@@ -1,0 +1,240 @@
+# Copyright 2023 The KerasNLP Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+"""Transformer encoder block implementation based on `keras.layers.Layer`."""
+
+from tensorflow import keras
+
+import keras_nlp
+from keras_nlp.api_export import keras_nlp_export
+from keras_nlp.utils.keras_utils import clone_initializer
+
+from keras_nlp.layers.transformer_layer_utils import (  # isort:skip
+    merge_padding_and_attention_mask,
+)
+
+
+@keras_nlp_export("keras_nlp.layers.XLNetEncoder")
+class XLNetEncoder(keras.layers.Layer):
+    """XLNet encoder.
+
+    This class follows the architecture of the transformer encoder layer in the
+    paper [Attention is All You Need](https://arxiv.org/abs/1706.03762). Users
+    can instantiate multiple instances of this class to stack up an encoder.
+
+    This layer will correctly compute an attention mask from an implicit
+    Keras padding mask (for example, by passing `mask_zero=True` to a
+    `keras.layers.Embedding` layer). See the Masking and Padding
+    [guide](https://keras.io/guides/understanding_masking_and_padding/)
+    for more details.
+
+    Args:
+        intermediate_dim: int, the hidden size of feedforward network.
+        num_heads: int, the number of heads in the
+            `keras.layers.MultiHeadRelativeAttention` layer.
+        dropout: float, defaults to 0. the dropout value, shared by
+            `keras.layers.MultiHeadRelativeAttention` and feedforward network.
+        activation: string or `keras.activations`, defaults to "relu". the
+            activation function of feedforward network.
+        layer_norm_epsilon: float, defaults to 1e-5. The epsilon value in layer
+            normalization components.
+        kernel_initializer: string or `keras.initializers` initializer,
+            defaults to "glorot_uniform". The kernel initializer for
+            the dense and multiheaded relative attention layers.
+        bias_initializer: string or `keras.initializers` initializer,
+            defaults to "zeros". The bias initializer for
+            the dense and multiheaded relative attention layers.
+        normalize_first: bool. Defaults to False. If True, the inputs to the
+            attention layer and the intermediate dense layer  are normalized
+            (similar to GPT-2). If set to False, outputs of attention layer and
+            intermediate dense layer are normalized (similar to XLNet).
+        name: string, defaults to None. The name of the layer.
+        **kwargs: other keyword arguments.
+
+    Examples:
+
+    ```python
+    # Create a single transformer encoder layer.
+    encoder = keras_nlp.layers.XLNetEncoder(
+        intermediate_dim=64, num_heads=8)
+
+    # Create a simple model containing the encoder.
+    input = keras.Input(shape=[10, 64])
+    output = encoder(input)
+    model = keras.Model(inputs=input, outputs=output)
+
+    # Call encoder on the inputs.
+    input_data = tf.random.uniform(shape=[2, 10, 64])
+    output = model(input_data)
+    ```
+
+    References:
+     - [](https://arxiv.org/abs/1906.08237)
+    """
+
+    def __init__(
+        self,
+        intermediate_dim,
+        num_heads,
+        dropout=0,
+        activation="relu",
+        layer_norm_epsilon=1e-05,
+        kernel_initializer="glorot_uniform",
+        bias_initializer="zeros",
+        normalize_first=False,
+        name=None,
+        **kwargs,
+    ):
+        # Work around for model saving
+        self._input_shape = kwargs.pop("build_input_shape", None)
+
+        super().__init__(name=name, **kwargs)
+        self.intermediate_dim = intermediate_dim
+        self.num_heads = num_heads
+        self.dropout = dropout
+        self.activation = keras.activations.get(activation)
+        self.layer_norm_epsilon = layer_norm_epsilon
+        self.kernel_initializer = keras.initializers.get(kernel_initializer)
+        self.bias_initializer = keras.initializers.get(bias_initializer)
+        self.normalize_first = normalize_first
+        self._built = False
+        self.supports_masking = True
+
+        if self._input_shape is not None:
+            self._build(self._input_shape)
+
+    def _build(self, input_shape):
+        # Create layers based on input shape.
+        self._built = True
+        self._input_shape = input_shape
+        # Infer the dimension of our hidden feature size from the build shape.
+        hidden_dim = input_shape[-1]
+        # Attention head size is `hidden_dim` over the number of heads.
+        key_dim = int(hidden_dim // self.num_heads)
+
+        # Relaive attention layers.
+        self._relative_attention_layer = (
+            keras_nlp.layers.TwoStreamRelativeAttention(
+                num_heads=self.num_heads,
+                key_dim=key_dim,
+                dropout=self.dropout,
+                kernel_initializer=clone_initializer(self.kernel_initializer),
+                bias_initializer=clone_initializer(self.bias_initializer),
+            )
+        )
+        self._relative_attention_layer._build_from_signature(
+            query=input_shape,
+            value=input_shape,
+        )
+        self._relative_attention_layernorm = keras.layers.LayerNormalization(
+            epsilon=self.layer_norm_epsilon,
+        )
+        self._relative_attention_dropout = keras.layers.Dropout(
+            rate=self.dropout,
+        )
+
+        # Feedforward layers.
+        self._feedforward_layernorm = keras.layers.LayerNormalization(
+            epsilon=self.layer_norm_epsilon,
+        )
+        self._feedforward_intermediate_dense = keras.layers.Dense(
+            self.intermediate_dim,
+            activation=self.activation,
+            kernel_initializer=clone_initializer(self.kernel_initializer),
+            bias_initializer=clone_initializer(self.bias_initializer),
+        )
+        self._feedforward_output_dense = keras.layers.Dense(
+            hidden_dim,
+            kernel_initializer=clone_initializer(self.kernel_initializer),
+            bias_initializer=clone_initializer(self.bias_initializer),
+        )
+        self._feedforward_dropout = keras.layers.Dropout(
+            rate=self.dropout,
+        )
+
+    def call(self, inputs, padding_mask=None, attention_mask=None):
+        """Forward pass of the XLNetEncoder.
+
+        Args:
+            inputs: a Tensor. The input data to XLNetEncoder, should be
+                of shape [batch_size, sequence_length, hidden_dim].
+            padding_mask: a boolean Tensor. It indicates if the token should be
+                masked because the token is introduced due to padding.
+                `padding_mask` should have shape [batch_size, sequence_length].
+            attention_mask: a boolean Tensor. Customized mask used to mask out
+                certain tokens. `attention_mask` should have shape
+                [batch_size, sequence_length, sequence_length].
+
+        Returns:
+            A Tensor of the same shape as the `inputs`.
+        """
+
+        if not self._built:
+            self._build(inputs.shape)
+
+        x = inputs  # Intermediate result.
+
+        # Compute self attention mask.
+        relative_attention_mask = merge_padding_and_attention_mask(
+            inputs, padding_mask, attention_mask
+        )
+
+        # Self attention block.
+        residual = x
+        if self.normalize_first:
+            x = self._relative_attention_layernorm(x)
+        x, y = self._relative_attention_layer(
+            query=x,
+            value=x,
+            attention_mask=relative_attention_mask,
+        )
+        x = self._relative_attention_dropout(x)
+        x = x + residual
+        if not self.normalize_first:
+            x = self._relative_attention_layernorm(x)
+
+        # Feedforward block.
+        residual = x
+        if self.normalize_first:
+            x = self._feedforward_layernorm(x)
+        x = self._feedforward_intermediate_dense(x)
+        x = self._feedforward_output_dense(x)
+        x = self._feedforward_dropout(x)
+        x = x + residual
+        if not self.normalize_first:
+            x = self._feedforward_layernorm(x)
+
+        return x
+
+    def get_config(self):
+        config = super().get_config()
+        config.update(
+            {
+                "intermediate_dim": self.intermediate_dim,
+                "num_heads": self.num_heads,
+                "dropout": self.dropout,
+                "activation": keras.activations.serialize(self.activation),
+                "layer_norm_epsilon": self.layer_norm_epsilon,
+                "kernel_initializer": keras.initializers.serialize(
+                    self.kernel_initializer
+                ),
+                "bias_initializer": keras.initializers.serialize(
+                    self.bias_initializer
+                ),
+                "normalize_first": self.normalize_first,
+                "build_input_shape": self._input_shape,
+            }
+        )
+        return config
